refactor: handle external and internal actions
