+++ __tests__/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import { workflowManagerReducer } from '../src';
+import { workflowManagerInternalReducer } from '../src';
 
 export const store = configureStore({
-  reducer: { workflowManagerReducer },
+  reducer: { workflowManagerInternalReducer },
 });
