+++ src/ducks/utils.ts
@@ -1,12 +1,12 @@
 import { createAction } from '@reduxjs/toolkit';
 
-import { prefix } from './workflowManager.slice';
+import { prefix } from '.';
 
 export const createWorkflowAction = (
   workflowAction: string,
   payload: Record<string, unknown>,
 ) => {
-  const actionName = `${prefix}/external/${workflowAction}`;
+  const actionName = `${prefix}/${workflowAction}`;
 
   return createAction<typeof payload>(actionName)(payload);
 };
